Better Ritual Stone Counter in Ritual Helper
Extended the existing EnumRuneType with "translationKey" and "patchouliColor" entries.

Reworked the RitualHelper's Rune Counter.  It uses a net.minecraft.util Tuple to return an integer (totalRunes) and a Map matching the EnumRuneType to that type's rune count.

Updated ItemRitualDiviner to use the new counting system.

Updated RitualInfoProcessor to use the new counting system.