Add Curios Support for Sigils
This was hilariously easy since the sigils operate on an Overwritten `inventoryTick` from the Item class.  Curios is already a ticking inventory.  I literally just had put the Sigil in a Curios slot and it works.

I picked the Charm slot for the sigils, and only tagged the sigils that have a constant activation use.
* Fast Miner
* Green Grove
* Magnetism
* Ice (Frost Walker, NYI)
* Sigil of Holding (since it can hold those active Sigils).